Give the debug remote config test a real flag set
IosDebugRemoteConfigTest built its production IosRemoteConfig with
flags = lazyOf(emptySet()). publishCurrent() only publishes keys present in
that set, so the remote value never reached RemoteConfigState and the three
tests that expect it read the default instead.

Move the TestRemoteFlag helper out of IosRemoteConfigTest, which already had
this right, into its own file so both tests share it.
