+++ core/feature-flags/implementation/src/iosTest/kotlin/com/thomaskioko/tvmaniac/featureflags/implementation/IosRemoteConfigTest.kt
@@ -3,13 +3,10 @@ package com.thomaskioko.tvmaniac.featureflags.implementation
 import app.cash.turbine.test
 import com.thomaskioko.tvmaniac.core.logger.fixture.FakeLogger
 import com.thomaskioko.tvmaniac.featureflags.FeatureFlag
-import com.thomaskioko.tvmaniac.featureflags.FeatureFlagsRemoteConfig
-import com.thomaskioko.tvmaniac.featureflags.RemoteFlag
 import com.thomaskioko.tvmaniac.featureflags.model.FeatureFlagFetchInterval
 import com.thomaskioko.tvmaniac.featureflags.testing.FakeRemoteConfigBridge
 import io.kotest.matchers.shouldBe
 import kotlinx.coroutines.test.runTest
-import kotlinx.datetime.LocalDate
 import kotlin.test.Test
 
 class IosRemoteConfigTest {
@@ -98,22 +95,4 @@ class IosRemoteConfigTest {
         flags = lazyOf(setOf(flag)),
         logger = FakeLogger(),
     )
-
-    private class TestRemoteFlag(key: String) : RemoteFlag(
-        key = key,
-        title = "Test",
-        description = "Test flag.",
-        dateAdded = LocalDate(2026, 1, 1),
-        defaultValue = false,
-        remote = NoOpRemoteConfig,
-    )
-
-    private object NoOpRemoteConfig : FeatureFlagsRemoteConfig {
-        override fun observeBoolean(key: String, default: Boolean) = kotlinx.coroutines.flow.flowOf(default)
-        override fun observeSource(key: String) = kotlinx.coroutines.flow.flowOf(
-            com.thomaskioko.tvmaniac.featureflags.model.FeatureFlagSource.Firebase,
-        )
-        override suspend fun refresh() = Unit
-        override suspend fun setDefaults(defaults: Map<String, Boolean>) = Unit
-    }
 }
